+++ src/alfasim_sdk/_internal/alfacase/case_description.py
@@ -2641,8 +2641,18 @@ class PhysicsDescription:
         default=None, validator=optional(instance_of(Path))
     )
     keep_former_results: bool = attr.ib(default=False, validator=instance_of(bool))
-    emulsion_model = attrib_enum(default=constants.EmulsionModelType.ModelDefault)
+    emulsion_model_enabled: bool = attr.ib(default=True, validator=instance_of(bool))
+    emulsion_relative_viscosity_model = attrib_enum(
+        default=constants.EmulsionRelativeViscosityModelType.ModelDefault
+    )
+    emulsion_droplet_size_model = attrib_enum(
+        default=constants.EmulsionDropletSizeModelType.ModelDefault
+    )
+    emulsion_inversion_point_model = attrib_enum(
+        default=constants.EmulsionInversionPointModelType.ModelDefault
+    )
     emulsion_model_plugin_id: str = attr.ib(default="", validator=instance_of(str))
+
     flash_model = attrib_enum(default=constants.FlashModel.HydrocarbonAndWater)
     correlations_package = attrib_enum(
         default=constants.CorrelationPackageType.Classical