+++ tests/common_testing/alfasim_sdk_common_testing/filled_case_descriptions.py
@@ -628,7 +628,10 @@ PHYSICS_DESCRIPTION = case_description.PhysicsDescription(
     restart_filepath=Path(__file__),
     keep_former_results=True,
     correlations_package=constants.CorrelationPackageType.Alfasim,
-    emulsion_model=constants.EmulsionModelType.Brauner2001,
+    emulsion_model_enabled=True,
+    emulsion_relative_viscosity_model=constants.EmulsionRelativeViscosityModelType.Brinkman1952,
+    emulsion_droplet_size_model=constants.EmulsionDropletSizeModelType.Brauner2001,
+    emulsion_inversion_point_model=constants.EmulsionInversionPointModelType.Brinkman1952AndYeh1964,
     flash_model=constants.FlashModel.HydrocarbonOnly,
 )